studymate/requirements.txt
streamlit>=1.36.0
pymupdf>=1.24.9
sentence-transformers>=3.0.1
faiss-cpu>=1.8.0.post1
numpy>=1.26.4
ibm-watsonx-ai>=1.1.12
python-dotenv>=1.0.1
pydantic>=2.8.2
tqdm>=4.66.4
